still working on having two processess at the same time

+++ Kernel/include/syscalls.h
@@ -34,7 +34,7 @@ int syscall_ioctl(unsigned int fd, unsigned long request, void * params);
 void syscall_pause(void);
 void syscall_halt(void);
 int syscall_exit(unsigned int code);
-int syscall_kill(int pid, int sig);
+int syscall_kill(uint64_t pid, int sig);
 int syscall_getpid(void);
 void* syscall_mmap(void* address, uint64_t size);
