+++ Kernel/syscall/syscalls.c
@@ -48,7 +48,7 @@ uint64_t int80h(uint64_t sysno, uint64_t RDI, uint64_t RSI, uint64_t RDX, uint64
 		break;
 
 		case SYSCALL_KILL:
-		exitno = syscall_kill((int) RDI, (int) RSI);
+		exitno = syscall_kill(RDI, (int) RSI);
 		break;
 
 		case SYSCALL_GETTIME:
